Show the search bar only when necessary.
This commit turns the logic upside down. Instead of hiding the search
bar when embedded maps is used, we hide the search bar initially and
show it if the map is not embedded. This way we are able to get rid of
a flash of the search bar in the embedded map.