Bugfix: don't clear empty search results.
Explanation: the name of the typeahead event typeahead:opened is a
bit misleading. It is called not only when the autosuggestions are shown but also
when the input field regains focus after the user has for example
responded to a native browser ui dialog like a geolocating permission dialog.
https://github.com/twitter/typeahead.js/issues/130#issuecomment-15687492

The event was used to clear overlapping search results from under
an autosuggest menu. However, the event was also fired when showing
a unit details view and the search input regained focus. Here the
resulting search_results collection reset caused the region to be
reset.

Solution: made the clearSearch command more robust so it
won't fire events when the search results are already empty.

Closes #108.